requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
asyncer==0.0.8
attrs==25.4.0
backoff==2.2.1
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.1
colorlog==6.10.1
diskcache==5.6.3
distro==1.9.0
dspy==3.0.3
et_xmlfile==2.0.0
fastuuid==0.14.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.9.0
gepa==0.0.7
google-auth==2.41.1
google-auth-oauthlib==1.2.2
gspread==6.2.1
gspread-formatting==1.2.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
json_repair==0.52.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
litellm==1.78.7
magicattr==0.1.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.0
numpy==2.3.4
oauthlib==3.3.1
openai==2.6.1
openpyxl==3.1.5
optuna==4.5.0
orjson==3.11.3
packaging==25.0
pandas==2.3.3
propcache==0.4.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
fastapi==0.115.6
python-socketio==5.11.4
uvicorn==0.34.0
python-multipart==0.0.20
